Conflicts:
	CMakeLists.txt
	build/cocos2d_libs.xcodeproj/project.pbxproj

+++ cocos/editor-support/cocosbuilder/CCBAnimationManager.cpp
@@ -609,7 +609,7 @@ ActionInterval* CCBAnimationManager::getEaseAction(ActionInterval *pAction, CCBK
     }
     else
     {
-        log("CCBReader: Unkown easing type %d", easingType);
+        log("CCBReader: Unkown easing type %d", static_cast<int>(easingType));
         return pAction;
     }
 }
